refactor(UserController): Add fromtoken route to UserController.
