+++ src/test/java/org/opentripplanner/middleware/controllers/api/UserControllerTest.java
@@ -1,5 +1,5 @@
 package org.opentripplanner.middleware.controllers.api;
 
-public class Auth0UserControllerTest {
+public class UserControllerTest {
 
 }
